delete category endpoint
