+++ routes/categories.js
@@ -1,0 +1,8 @@
+//controllers
+const { deleteCategory } = require("../controllers/categories");
+
+const router = require("express").Router();
+
+router.delete("/:id", deleteCategory);
+
+module.exports = router;
