+++ routes/index.js
@@ -1,9 +1,10 @@
-const express = require('express')
-const usersRouter = require('./users')
+const express = require("express");
 
-const router = express.Router()
+const routerCategories = require("./categories");
+
+const router = express.Router();
 
 // example of a route with index controller get function
-router.use('/users', usersRouter)
+router.use("/categories", routerCategories);
 
-module.exports = router
+module.exports = router;
